Change onWakeUp() to onAttachView(TiView) and onSleep() to onDetachView()
The new naming is more convenient thoughout most android MPV libraries. Switching to Ti should now be simpler. It also reflects better than before that the it's only the view being absent. wakeup and sleep could imply the Presenter is kind of starting/stopping which is not the case.

onAttachView(TiView) has the attached view as first parameter which is @NonNull. No calls to getView() are neccessary inside onAttachView(TiView).
getView() is now @Nullable (and always was null when not between wakeup() and sleep() state) and causes lint to hint a Nullable warning. This warning was missing and caused a lot of NPE for Ti beginners