Implement TiFragmentDelegate analog to TiActivityDelegate
It's not possible to pass a single Object into the delegate because in CompositeAndroid the TiViewProvider and the Fragment are different objects. Adding all interfaces one by one helps when testing. There is no god object required which implements all interfaces at once.

Logging tag and toString belong in the Fragments itself rather than the delegate. Before it was using the logging tag "TiFragmentDelegate" instead of "SampleFragment" which doesn't help a lot when debugging

provideView was moved into the Fragments because this works differently between Compositeandroid and the Ti implementations