Fix crash when calling sendToView while the Presenter attaches the view
It was possible to call sendToView when the view set but a uiThreadExecutor was not.

// attachView(view)
       ...
        mView = view;
        moveToState(State.VIEW_ATTACHED, false);
       ...

sendToView was called after `mView = view;` on a different Thread.

This fix checks both (view and mUiThreadExecutor) to be available before calling `runOnUiThread`.

Additionally `sendPostponedActionsToView` will now be called when setting a new  uiThreadExecutor in case a third party Presenter host has to change the uiThreadExecutor while the view is attached or attaches it after onAttachView() for some reason